# tiki-backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vona API",
    "version": "1.0.0",
    "description": "API documentation for Vona application"
  },
  "servers": [
    {
      "url": "http://localhost:3003/api/v1",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/users": {
      "get": {
        "summary": "Get list of users",
        "description": "Retrieve a paginated list of users with optional filters (requires user login)",
        "operationId": "getAllUsers",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "description": "Filter parameters",
            "schema": { "$ref": "#/components/schemas/UserFilter" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "description": "Create a new user with provided data",
        "operationId": "createUser",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserCreateInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/users/user": {
      "get": {
        "summary": "Get current user",
        "description": "Retrieve the details of the currently authenticated user",
        "operationId": "getUserByOne",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "summary": "Get user by ID",
        "description": "Retrieve a user by their UUID (requires admin login)",
        "operationId": "getUserById",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "UUID of the user",
            "schema": { "$ref": "#/components/schemas/UserIdParam" }
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a user",
        "description": "Update user details by UUID (requires user login)",
        "operationId": "updateUser",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "UUID of the user",
            "schema": { "$ref": "#/components/schemas/UserIdParam" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserUpdateInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "description": "Delete a user by UUID (requires user login)",
        "operationId": "deleteUser",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "UUID of the user",
            "schema": { "$ref": "#/components/schemas/UserIdParam" }
          }
        ],
        "responses": {
          "204": { "description": "User deleted successfully" },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/users/upload-avatar/{userId}": {
      "post": {
        "summary": "Upload user avatar",
        "description": "Upload an avatar image for a user by UUID (requires user login)",
        "operationId": "uploadAvatar",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "UUID of the user",
            "schema": { "$ref": "#/components/schemas/UserIdParam" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary",
                    "description": "Avatar image file"
                  }
                },
                "required": ["avatar"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avatar uploaded successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid file or input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/auth/logout-all": {
      "post": {
        "summary": "Logout from all devices",
        "description": "Log out the user from all devices (invalidates all tokens)",
        "operationId": "logoutAllDevices",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logout from all devices successful" },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/refresh-token": {
      "post": {
        "summary": "Refresh JWT token",
        "description": "Generate a new JWT token using a valid refresh token",
        "operationId": "refreshToken",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "New JWT token generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "New JWT token"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "summary": "Request password reset",
        "description": "Send a password reset link to the user’s email",
        "operationId": "forgotPassword",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ForgotPasswordRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset link sent",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password/{userId}/{token}": {
      "post": {
        "summary": "Reset password",
        "description": "Reset user password using a reset token",
        "operationId": "resetPassword",
        "tags": ["Auth"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "UUID of the user",
            "schema": { "$ref": "#/components/schemas/UserIdParam" }
          },
          {
            "name": "token",
            "in": "path",
            "required": true,
            "description": "Password reset token",
            "schema": { "$ref": "#/components/schemas/ResetPasswordTokenParam" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResetPasswordRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input or token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/enable-2fa": {
      "post": {
        "summary": "Enable two-factor authentication",
        "description": "Enable 2FA for the authenticated user",
        "operationId": "enableTwoFactor",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "2FA enabled successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/verify-2fa": {
      "post": {
        "summary": "Verify two-factor authentication",
        "description": "Verify 2FA code to complete login",
        "operationId": "verifyTwoFactor",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TwoFactorRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "2FA verification successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid 2FA code",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserResponse": {
        "type": "object",
        "properties": {
          "uuid": { "type": "string", "format": "uuid" },
          "username": { "type": "string" },
          "full_name": { "type": "string", "nullable": true },
          "email": { "type": "string", "format": "email" },
          "phone_number": { "type": "string", "nullable": true },
          "day_of_birth": { "type": "string", "nullable": true },
          "is_public": { "type": "boolean" },
          "is_deleted": { "type": "boolean" },
          "gender": { "type": "boolean", "nullable": true },
          "role_id": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "avatar": { "type": "string", "nullable": true },
          "created_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          },
          "updated_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          },
          "addresses": { "type": "array", "items": {} },
          "orders": { "type": "array", "items": {} },
          "cart_items": { "type": "array", "items": {} },
          "reviews": { "type": "array", "items": {} },
          "conversations": { "type": "array", "items": {} }
        },
        "required": [
          "uuid",
          "username",
          "full_name",
          "email",
          "phone_number",
          "day_of_birth",
          "is_public",
          "is_deleted",
          "gender",
          "role_id",
          "avatar",
          "created_at",
          "updated_at"
        ]
      },
      "UserCreateInput": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "minLength": 3 },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "full_name": { "type": "string", "nullable": true },
          "phone_number": { "type": "string", "nullable": true },
          "day_of_birth": { "type": "string", "nullable": true },
          "gender": { "type": "boolean" },
          "avatar": { "type": "string" },
          "role_id": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "default": 1
          },
          "is_public": { "type": "boolean", "default": true },
          "is_activated": { "type": "boolean", "default": true }
        },
        "required": [
          "username",
          "email",
          "password",
          "full_name",
          "phone_number",
          "day_of_birth",
          "role_id",
          "is_public",
          "is_activated"
        ]
      },
      "UserUpdateInput": {
        "type": "object",
        "properties": {
          "two_factor_enabled": { "type": "boolean" },
          "username": { "type": "string", "minLength": 3 },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "full_name": { "type": "string", "nullable": true },
          "phone_number": { "type": "string", "nullable": true },
          "day_of_birth": { "type": "string", "nullable": true },
          "gender": { "type": "boolean" },
          "avatar": { "type": "string" },
          "role_id": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "default": 1
          },
          "is_public": { "type": "boolean", "default": true },
          "is_activated": { "type": "boolean", "default": true }
        },
        "required": [
          "username",
          "email",
          "password",
          "full_name",
          "phone_number",
          "day_of_birth",
          "role_id",
          "is_public",
          "is_activated"
        ]
      },
      "UserFilter": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "username": { "type": "string", "minLength": 3 },
          "full_name": { "type": "string" },
          "phone_number": { "type": "string" },
          "sort_by_name": { "type": "string", "enum": ["asc", "desc"] },
          "sort_by_date": { "type": "string", "enum": ["newest", "oldest"] },
          "limit": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "offset": { "type": "integer", "minimum": 0 }
        }
      },
...
      "ForgotPasswordRequest": {
        "type": "object",
        "properties": { "email": { "type": "string", "format": "email" } },
        "required": ["email"]
      },
      "ResetPasswordRequest": {
        "type": "object",
        "properties": { "password": { "type": "string", "minLength": 6 } },
        "required": ["password"]
      },
      "TwoFactorRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "twoFactorCode": { "type": "string", "minLength": 6, "maxLength": 6 }
        },
        "required": ["email", "twoFactorCode"]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "twoFactorRequired": { "type": "boolean" },
          "uuid": { "type": "string", "format": "uuid" },
          "username": { "type": "string" },
          "full_name": { "type": "string", "nullable": true },
          "email": { "type": "string", "format": "email" },
          "phone_number": { "type": "string", "nullable": true },
          "day_of_birth": { "type": "string", "nullable": true },
          "is_public": { "type": "boolean" },
          "is_deleted": { "type": "boolean" },
          "gender": { "type": "boolean", "nullable": true },
          "role_id": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "avatar": { "type": "string", "nullable": true },
          "created_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          },
          "updated_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          },
          "addresses": { "type": "array", "items": {} },
          "orders": { "type": "array", "items": {} },
          "cart_items": { "type": "array", "items": {} },
          "reviews": { "type": "array", "items": {} },
          "conversations": { "type": "array", "items": {} }
        },
        "required": [
          "uuid",
          "username",
          "full_name",
          "email",
          "phone_number",
          "day_of_birth",
          "is_public",
          "is_deleted",
          "gender",
          "role_id",
          "avatar",
          "created_at",
          "updated_at"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "data": { "nullable": true },
          "statusCode": { "type": "number" }
        },
        "required": ["error", "statusCode"]
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "data": {},
          "statusCode": { "type": "number" }
        },
        "required": ["message", "statusCode"]
      },
      "UserIdParam": {
        "type": "object",
        "properties": { "userId": { "type": "string", "format": "uuid" } },
        "required": ["userId"]
      },
      "ResetPasswordTokenParam": {
        "type": "object",
        "properties": { "token": { "type": "string" } },
        "required": ["token"]
      },
...
    }
  }
}